refactor: move non-useCases to `domain` package
