+++ app/src/main/kotlin/com/wire/android/feature/conversation/content/domain/SendMessageService.kt
@@ -1,4 +1,4 @@
-package com.wire.android.feature.conversation.content.usecase
+package com.wire.android.feature.conversation.content.domain
 
 import com.wire.android.core.exception.Failure
 import com.wire.android.core.exception.NetworkConnection
@@ -6,6 +6,7 @@ import com.wire.android.core.functional.Either
 import com.wire.android.core.functional.suspending
 import com.wire.android.feature.conversation.content.Message
 import com.wire.android.feature.conversation.content.MessageRepository
+import com.wire.android.feature.conversation.content.usecase.SendMessageWorkerScheduler
 
 class SendMessageService(
     private val messageRepository: MessageRepository,
