+++ app/src/test/kotlin/com/wire/android/feature/conversation/content/domain/OutgoingMessageRecipientsRetrieverTest.kt
@@ -8,7 +8,6 @@ import com.wire.android.feature.contact.Contact
 import com.wire.android.feature.contact.ContactClient
 import com.wire.android.feature.contact.DetailedContact
 import com.wire.android.feature.conversation.content.MessageRepository
-import com.wire.android.feature.conversation.content.domain.OutgoingMessageRecipientsRetriever
 import com.wire.android.feature.conversation.data.ConversationRepository
 import com.wire.android.framework.functional.shouldFail
 import com.wire.android.framework.functional.shouldSucceed
